+++ src/App/VRChatContentPublisher.App/Views/MainWindow.axaml.cs
@@ -3,6 +3,7 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Interactivity;
 using Avalonia.Reactive;
+using VRChatContentPublisher.App.Interop.Windows;
 using VRChatContentPublisher.App.ViewModels;
 
 namespace VRChatContentPublisher.App.Views;
@@ -33,9 +34,7 @@ public partial class MainWindow : Window
         {
             if (DataContext is MainWindowViewModel viewModel && (viewModel.Pinned || !viewModel.Borderless))
                 return;
-#if !DEBUG
             Hide();
-#endif
         };
 
         _lastStateBeforeMinimized = WindowState;
@@ -47,6 +46,14 @@ public partial class MainWindow : Window
 
                 _lastStateBeforeMinimized = state;
             }));
+
+        if (OperatingSystem.IsWindows())
+        {
+#pragma warning disable CA1416
+            // https://github.com/AvaloniaUI/Avalonia/issues/19006#issuecomment-3562753006
+            Win32Properties.AddWorkingAreaChangedListener(this, UpdateWindowConfiguration);
+#pragma warning restore CA1416
+        }
     }
 
     private bool IsBorderlessSupported() => OperatingSystem.IsWindows();
@@ -84,80 +91,43 @@ public partial class MainWindow : Window
 
     private void UpdateWindowPosition()
     {
+        if (!OperatingSystem.IsWindows())
+            return;
         if (Screens.Primary is not { } primaryScreen)
             return;
 
-        var taskBarLocation = GetTaskBarLocation();
-        var taskBarHeight = GetTaskBarHeight();
+        var workingArea = WorkingAreaHelper.GetWorkingArea();
+        var taskBarLocation = TaskBarHelper.GetTaskBarLocation(primaryScreen, workingArea);
+        var taskBarHeight = TaskBarHelper.GetTaskBarHeight(primaryScreen, workingArea);
 
         var screenBoundsWithDpi = new PixelPoint(primaryScreen.Bounds.Width, primaryScreen.Bounds.Height);
         var windowBoundsWithDpi =
             PixelPoint.FromPoint(new Point((int)Bounds.Size.Width, (int)Bounds.Size.Height), primaryScreen.Scaling);
         switch (taskBarLocation)
         {
-            case TaskBarLocation.Top:
+            case TaskBarHelper.TaskBarLocation.Top:
+                Position = screenBoundsWithDpi.WithY(0) - windowBoundsWithDpi.WithY(0) + 
+                           PixelPoint.FromPoint(new Point(0, taskBarHeight), primaryScreen.Scaling) -
+                           PixelPoint.FromPoint(new Point(16, -12), primaryScreen.Scaling);
                 break;
-            case TaskBarLocation.Bottom:
+            case TaskBarHelper.TaskBarLocation.Bottom:
                 Position = screenBoundsWithDpi - windowBoundsWithDpi -
                            PixelPoint.FromPoint(new Point(0, taskBarHeight), primaryScreen.Scaling) -
                            PixelPoint.FromPoint(new Point(16, 12), primaryScreen.Scaling);
                 break;
-            case TaskBarLocation.Left:
+            case TaskBarHelper.TaskBarLocation.Left:
+                Position = screenBoundsWithDpi.WithX(0) - windowBoundsWithDpi.WithX(0) +
+                           PixelPoint.FromPoint(new Point(taskBarHeight, 0), primaryScreen.Scaling) +
+                           PixelPoint.FromPoint(new Point(0, -12), primaryScreen.Scaling);
                 break;
-            case TaskBarLocation.Right:
+            case TaskBarHelper.TaskBarLocation.Right:
+                Position = screenBoundsWithDpi - windowBoundsWithDpi -
+                           PixelPoint.FromPoint(new Point(taskBarHeight, 0), primaryScreen.Scaling) -
+                           PixelPoint.FromPoint(new Point(16, 12), primaryScreen.Scaling);
                 break;
             default:
                 throw new ArgumentOutOfRangeException();
         }
-    }
-
-    #endregion
-
-    #region TaskBar Helper
-
-    private enum TaskBarLocation
-    {
-        Top,
-        Bottom,
-        Left,
-        Right,
-        Unknown
-    }
-
-    private TaskBarLocation GetTaskBarLocation()
-    {
-        if (Screens.Primary is not { } primaryScreen)
-            return TaskBarLocation.Unknown;
-
-        var taskBarOnTopOrBottom = primaryScreen.WorkingArea.Width == primaryScreen.Bounds.Width;
-        if (taskBarOnTopOrBottom)
-        {
-            if (primaryScreen.WorkingArea.TopLeft.Y > 0)
-                return TaskBarLocation.Top;
-        }
-        else
-        {
-            return primaryScreen.WorkingArea.TopLeft.X > 0 ? TaskBarLocation.Left : TaskBarLocation.Right;
-        }
-
-        return TaskBarLocation.Bottom;
-    }
-
-    private int GetTaskBarHeight()
-    {
-        if (Screens.Primary is not { } primaryScreen)
-            return -1;
-
-        var taskbarHeight = GetTaskBarLocation() switch
-        {
-            TaskBarLocation.Top => primaryScreen.WorkingArea.Height,
-            TaskBarLocation.Bottom => primaryScreen.Bounds.Height - primaryScreen.WorkingArea.Height,
-            TaskBarLocation.Left => primaryScreen.Bounds.Width - primaryScreen.WorkingArea.Width,
-            TaskBarLocation.Right => primaryScreen.Bounds.Height - primaryScreen.WorkingArea.Height,
-            _ => throw new ArgumentOutOfRangeException()
-        };
-
-        return (int)(taskbarHeight / primaryScreen.Scaling);
     }
 
     #endregion
@@ -199,5 +169,6 @@ public partial class MainWindow : Window
         Activate();
 
         WindowState = _lastStateBeforeMinimized;
+        if (WindowState == WindowState.Normal) UpdateWindowConfiguration();
     }
 }